correct

+++ Termux/crdl.sh
@@ -231,15 +231,17 @@ crdlAccessTime=$(stat -c "%x" $FIRST_INSTALL | awk '{print $1, substr($2,1,5)}')
 currentTime=$(date "+%Y-%m-%d %H:%M")
 
 # --- Shizuku Setup first time ---
-if ! $HOME/rish -c "id" >/dev/null 2>&1 && [ ! -f "$LAST_INSTALL" ] && [ "crdlAccessTime" == "currentTime" ]; then
+if ! $HOME/rish -c "id" >/dev/null 2>&1 && ! su -c "id" >/dev/null 2>&1 && [ ! -f "$LAST_INSTALL" ] && [ "$crdlAccessTime" == "$currentTime" ]; then
+  echo -e "$info Please manually install Shizuku from Google Play Store." && sleep 1
   termux-open-url "https://play.google.com/store/apps/details?id=moe.shizuku.privileged.api"
   if [ ! -f "$HOME/rish" ] || [ ! -f "$HOME/rish_shizuku.dex" ]; then
-    curl -o "$HOME/rish" "https://raw.githubusercontent.com/arghya339/crdl/refs/heads/main/Termux/Shizuku/rish" && chmod +x "$HOME/rish" > /dev/null 2>&1
+    curl -o "$HOME/rish" "https://raw.githubusercontent.com/arghya339/crdl/refs/heads/main/Termux/Shizuku/rish" > /dev/null 2>&1 && chmod +x "$HOME/rish"
     sleep 0.5 && curl -o "$HOME/rish_shizuku.dex" "https://raw.githubusercontent.com/arghya339/crdl/refs/heads/main/Termux/Shizuku/rish_shizuku.dex" > /dev/null 2>&1
   fi
+  echo -e "$info Please start Shizuku by following guide. Then rerun script by running ~ ${Green}crdl${Reset}" && sleep 1
   if [ $Android -le 10 ]; then
     termux-open-url "https://youtu.be/ZxjelegpTLA"  # YouTube/@MrPalash360: Start Shizuku using Computer
-  elif [ $Android -ge 11 ]; then
+  else
     termux-open-url "https://youtu.be/YRd0FBfdntQ"  # YouTube/@MrPalash360: Start Shizuku Android 11+
   fi
   exit 1
